# backend/openapi.yaml
openapi: 3.0.3
info:
  title: ITFY E-Voting Platform API
  description: |
    # ITFY E-Voting Platform Backend API
    
    Welcome to the comprehensive API documentation for the ITFY E-Voting platform.
    This platform enables event organizers to create and manage voting events with 
    candidates, categories, and secure payment-based voting.
    
    ## Features
    - **Authentication & Authorization**: Secure JWT-based authentication with role-based access control
    - **Event Management**: Create, manage, and publish voting events
    - **Category Management**: Organize candidates into voting categories
    - **Candidate Management**: Register and manage candidates with profiles
    - **Voting System**: Purchase vote bundles and cast votes securely
    - **Payment Integration**: Paystack payment gateway integration
    - **Forms & Submissions**: Dynamic nomination forms with approval workflows
    - **Analytics**: Comprehensive voting analytics and reporting
    
    ## Authentication
    Most endpoints require authentication via JWT tokens. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your_token>
    ```
    
    ## User Roles
    - **super_admin**: Full system access
    - **admin**: Administrative access to events and users
    - **organiser**: Can create and manage events
    - **moderator**: Can view and moderate content
    
    ## Rate Limiting
    API requests are rate-limited to prevent abuse:
...
    
    ## Pagination
...
    
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
    CandidateAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authenticated candidates
  schemas:
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "An error occurred"
        error:
          type: string
          example: "Detailed error message"
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation successful"
        data:
          type: object
...
    ObjectId:
      type: string
      pattern: '^[a-fA-F0-9]{24}$'
      example: "507f1f77bcf86cd799439011"

    # ==================
    # User Schemas
    # ==================
    User:
      type: object
      properties:
        _id:
          $ref: '#/components/schemas/ObjectId'
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "user@example.com"
        role:
          type: string
          enum: [super_admin, admin, organiser, moderator]
          example: "admin"
        permissions:
          type: array
          items:
            type: string
            enum: [read, write, update, delete, super]
          example: ["read", "write"]
        email_verified:
          type: boolean
          example: true
        email_verified_at:
          type: string
          format: date-time
          nullable: true
        last_login:
          type: string
          format: date-time
          nullable: true
        bio:
          type: string
          nullable: true
        photo_url:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive, suspended, deleted]
          example: "active"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # ==================
    # Auth Request Schemas
    # ==================
...
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Login successful"
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            tokens:
              type: object
              properties:
                accessToken:
                  type: string
                  example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                refreshToken:
                  type: string
                  example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                expiresIn:
                  type: string
                  example: "15m"
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
    ForgotPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
    ResetPasswordRequest:
      type: object
      required:
        - token
        - password
      properties:
        token:
          type: string
          example: "reset-token-here"
        password:
          type: string
          format: password
          minLength: 8
          example: "NewSecurePass123!"
    ChangePasswordRequest:
      type: object
      required:
        - currentPassword
        - newPassword
      properties:
        currentPassword:
          type: string
          format: password
          example: "CurrentPass123!"
        newPassword:
          type: string
          format: password
          minLength: 8
          example: "NewSecurePass123!"
    VerifyEmailRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          example: "verification-token-here"
    ResendVerificationRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"

    # ==================
    # Candidate Auth Schemas
    # ==================
    CandidateLoginRequest:
      type: object
      required:
        - identifier
        - password
      properties:
        identifier:
          type: string
          description: Candidate code or email
          example: "CAN001"
        password:
          type: string
          format: password
          example: "CandidatePass123!"

    # ==================
    # Event Schemas
    # ==================
...
  responses:
    UnauthorizedError:
      description: Authentication token is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: "Unauthorized"
            error: "You are not authorized to perform this action"
...
    NotFoundError:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: "Not Found"
            error: "The requested resource was not found"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: "Validation Error"
            error: "Some required fields are missing or invalid"
...
paths:
...
  /auth/refresh-token:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Get a new access token using a valid refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      accessToken:
                        type: string
                      expiresIn:
                        type: string
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/verify-email:
    post:
      tags:
        - Authentication
      summary: Verify email address
      description: Verify user's email using verification token
      operationId: verifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
      responses:
        '200':
          description: Email verified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Invalid or expired token
  /auth/resend-verification:
    post:
      tags:
        - Authentication
      summary: Resend verification email
      description: Resend email verification link
      operationId: resendVerification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResendVerificationRequest'
      responses:
        '200':
          description: Verification email sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /auth/forgot-password:
    post:
      tags:
        - Authentication
      summary: Request password reset
      description: Send password reset email to user
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Password reset email sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /auth/reset-password:
    post:
      tags:
        - Authentication
      summary: Reset password
      description: Reset user password using reset token
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: Invalid or expired reset token
  /auth/change-password:
    post:
      tags:
        - Authentication
      summary: Change password
      description: Change the authenticated user's password
      operationId: changePassword
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Current password is incorrect
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user
      description: Retrieve the authenticated user's profile
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # ==================
  # Candidate Authentication
  # ==================
  /auth/candidate/login:
    post:
      tags:
        - Authentication
      summary: Login as candidate
      description: Authenticate candidate with code/email and password
      operationId: candidateLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CandidateLoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
  /auth/candidate/logout:
    post:
      tags:
        - Authentication
      summary: Logout candidate
      description: Invalidate candidate tokens
      operationId: candidateLogout
      security:
        - CandidateAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /auth/candidate/forgot-password:
    post:
      tags:
        - Authentication
      summary: Request candidate password reset
      description: Send password reset email to candidate
      operationId: candidateForgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Password reset email sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /auth/candidate/reset-password:
    post:
      tags:
        - Authentication
      summary: Reset candidate password
      description: Reset candidate password using token
      operationId: candidateResetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /auth/candidate/me:
    get:
      tags:
        - Authentication
      summary: Get current candidate
      description: Retrieve authenticated candidate's profile
      operationId: candidateMe
      security:
        - CandidateAuth: []
      responses:
        '200':
          description: Candidate profile retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # ==================
  # User Endpoints
  # ==================
...